Break out meta info in the menu buttons to its own file
