+++ src/components/app/MenuButtons/index.css
@@ -185,48 +185,3 @@
   bottom: 1px;
   vertical-align: middle;
 }
-
-.menuButtonsOpenMetaInfoButtonBox,
-.menuButtonsOpenMetaInfoButtonLabel,
-.menuButtonsOpenMetaInfoButtonButton {
-  height: 100%;
-  font-size: 100%;
-}
-
-.menuButtonsOpenMetaInfoButtonBox {
-  display: flex;
-
-  /* Let this box shrink down to the size of the three dots and a bit of text. */
-  min-width: 75px;
-
-  /* Shrink down twice as fast as the range selection labels. */
-  flex-shrink: 2;
-  padding: 0;
-}
-
-.menuButtonsOpenMetaInfoButtonLabel {
-  position: relative;
-  display: flex;
-
-  /* Let this box shrink down to the size of the three dots. */
-  min-width: 0;
-
-  /* Make sure the padding doesn't contribute to the height, which is 100% */
-  box-sizing: border-box;
-  flex-flow: column nowrap;
-  align-items: stretch;
-  padding: 5px 10px;
-}
-
-.menuButtonsOpenMetaInfoButtonLabelOverflow {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.menuButtonsOpenMetaInfoButtonBox .buttonWithPanelButton,
-.menuButtonsOpenMetaInfoButtonBox .buttonWithPanel:hover,
-.menuButtonsOpenMetaInfoButtonBox .buttonWithPanel.open {
-  width: 40px;
-  background: url(../../../../res/img/svg/meatball.svg) 50% 50% no-repeat;
-}
